fix: 🐛 move test to wechaty-base
